Updates tests related to EvdSocket to reflect new on-read notification policy
The new policy is, basically, that if a read operation returns as many bytes as
requested, is application's responsability to schedule and perform read operation
again; as no new on-read notification will be launched. Applications should only
assume that there is no more data to read when a read operation returned less bytes
than requested; in which case a new on-read event will be fired by the framework as
soon as new data is ready to be read.