Makes EvdWebTransport to be the transport associated to peers
instead of EvdLongPolling of EvdWebSocket. The actual mechanism transport is
then associated to the peer using a GObject data value.

This brings consistency in the sense that a peer coming from a web transport now
has that web transport as the value of its 'transport' property, instead of
the underlaying mechanism actually used (longpolling/websocket transport).