Deprecates set_selector() and get_selector() of EvdWebTransportServer
From now on, the selector is not a property of the transport. Instead,
the transport can use one or more selectors as entry points, by using
the newly added API use_selector() and unuse_selector().

In future releases, set_selector(), get_selector() and the 'selector'
property will be removed completely.