Makes EvdIoStreamGroup to watch 'close' signal of streams and auto-remove them
This implies a significant change in how streams and groups are handled. Before,
specialized IO stream groups like EvdService and EvdConnectionPool were
responsible for removing the streams when the were closed. Now the base IO
stream group handles that automatically by default.

Stream groups can still override the add() or remove() virtual methods to
implement a different logic, and can chain up to EvdIoStreamGroup's base
methods to use this new stream auto-tracking feature.

This change also implies that EvdIoStreamGroup now maintains a GList with its
streams to be able to disconnect from 'close' signal when the group is disposed.