Removes IO stream tracking in EvdIoStreamGroup
This turned out not to be such a good idea. Maintaining a GList of streams
implies iterating the list every time an element is removed from the group.
This is not efficient performance-wise.

Now, it is the EvdIoStream object the one responsible to remove itself from
its group (if any) when closed or destroyed.